requirements.txt
# ME-Benchmark requirements
torch>=1.13.0
transformers>=4.30.0
datasets>=2.14.0
accelerate>=0.20.0
pyyaml>=6.0
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
tqdm>=4.64.0
matplotlib>=3.5.0
seaborn>=0.11.0